+++ bellows/src/uefi.rs
@@ -1,0 +1,107 @@
+// bellows/src/uefi.rs
+
+use alloc::vec::Vec;
+use core::ffi::c_void;
+
+/// A simple Result type for our bootloader,
+/// returning a static string on error.
+pub type Result<T> = core::result::Result<T, &'static str>;
+
+/// Minimal subset of UEFI memory types (only those we need)
+#[repr(usize)]
+pub enum EfiMemoryType {
+    EfiLoaderData = 2,
+    EfiMaxMemoryType = 15,
+}
+
+/// Minimal UEFI System Table and protocols used by this loader
+#[repr(C)]
+pub struct EfiSystemTable {
+    _hdr: [u8; 24],
+    pub con_out: *mut EfiSimpleTextOutput,
+    _con_in: *mut c_void,
+    pub boot_services: *mut EfiBootServices,
+}
+
+/// Very small subset of Boot Services we call
+#[repr(C)]
+pub struct EfiBootServices {
+    _pad: [usize; 24],
+    /// allocate_pages(AllocateType, MemoryType, Pages, *mut PhysicalAddress) -> EFI_STATUS
+    pub allocate_pages: extern "efiapi" fn(usize, EfiMemoryType, usize, *mut usize) -> usize,
+    /// free_pages(PhysicalAddress, Pages) -> EFI_STATUS
+    pub free_pages: extern "efiapi" fn(usize, usize) -> usize,
+    _pad_2: [usize; 4],
+    /// locate_protocol(ProtocolGUID, Registration, *mut *Interface) -> EFI_STATUS
+    pub locate_protocol: extern "efiapi" fn(*const u8, *mut c_void, *mut *mut c_void) -> usize,
+    _pad_3: [usize; 3],
+    /// get_memory_map(MemoryMapSize, *MemoryMap, *MapKey, *DescriptorSize, *DescriptorVersion) -> EFI_STATUS
+    pub get_memory_map:
+        extern "efiapi" fn(*mut usize, *mut c_void, *mut usize, *mut usize, *mut u32) -> usize,
+    _pad_4: [usize; 2],
+    /// exit_boot_services(ImageHandle, MapKey) -> EFI_STATUS
+    pub exit_boot_services: extern "efiapi" fn(usize, usize) -> usize,
+}
+
+/// SimpleTextOutput protocol (we only use OutputString)
+#[repr(C)]
+pub struct EfiSimpleTextOutput {
+    _pad: [usize; 4], // skip many fields; we only use output_string
+    pub output_string: extern "efiapi" fn(*mut EfiSimpleTextOutput, *const u16) -> usize,
+}
+
+/// Simple FileSystem and File prototypes (very small subset)
+#[repr(C)]
+pub struct EfiSimpleFileSystem {
+    _revision: u64,
+    pub open_volume: extern "efiapi" fn(*mut EfiSimpleFileSystem, *mut *mut EfiFile) -> usize,
+}
+
+/// GUID for EFI_FILE_INFO protocol
+pub const EFI_FILE_INFO_GUID: [u8; 16] = [
+    0x0d, 0x95, 0xde, 0x05, 0x93, 0x31, 0xd2, 0x11, 0x8a, 0x41, 0x00, 0xa0, 0xc9, 0x3e, 0xc7, 0xea,
+];
+
+/// GUID for EFI_SIMPLE_FILE_SYSTEM_PROTOCOL
+pub const EFI_SIMPLE_FILE_SYSTEM_PROTOCOL_GUID: [u8; 16] = [
+    0x22, 0x5b, 0x4e, 0x96, 0x59, 0x64, 0xd2, 0x11, 0x8e, 0x39, 0x00, 0xa0, 0xc9, 0x69, 0x72, 0x3b,
+];
+
+#[repr(C)]
+pub struct EfiFile {
+    _revision: u64,
+    pub open: extern "efiapi" fn(*mut EfiFile, *mut *mut EfiFile, *const u16, u64, u64) -> usize,
+    pub close: extern "efiapi" fn(*mut EfiFile) -> usize,
+    _delete: extern "efiapi" fn(*mut EfiFile) -> usize,
+    pub read: extern "efiapi" fn(*mut EfiFile, *mut u64, *mut u8) -> usize,
+    _write: extern "efiapi" fn(*mut EfiFile, *mut u64, *mut u8) -> usize,
+    _reserved: usize,
+    _get_position: extern "efiapi" fn(*mut EfiFile, *mut u64) -> usize,
+    _set_position: extern "efiapi" fn(*mut EfiFile, u64) -> usize,
+    pub get_info: extern "efiapi" fn(*mut EfiFile, *const u8, *mut usize, *mut c_void) -> usize,
+    _set_info: extern "efiapi" fn(*mut EfiFile, *const u8, usize, *mut c_void) -> usize,
+    pub flush: extern "efiapi" fn(*mut EfiFile) -> usize,
+}
+
+#[repr(C)]
+pub struct EfiFileInfo {
+    _size: u64,
+    pub file_size: u64,
+    _physical_size: u64,
+    _create_time: u64,
+    _last_access_time: u64,
+    _modification_time: u64,
+    _attribute: u64,
+    _file_name: [u16; 1],
+}
+
+/// Print a &str to the UEFI console via SimpleTextOutput (OutputString)
+pub fn uefi_print(st: &EfiSystemTable, s: &str) {
+    let mut ucs2: Vec<u16> = s.encode_utf16().collect();
+    ucs2.push(0);
+    unsafe {
+        if !st.con_out.is_null() {
+            ((*st.con_out).output_string)(st.con_out, ucs2.as_ptr());
+        }
+    }
+}
